Fake a return value for "empty" input.
This only happens when the user enters non-empty input, but the code cleaner cleans everything up and the resulting code is an empty string. It turns out evaling an empty string in PHP returns false, which is not the return value users expect from "no-op" input.

One great example is `use` statements, which are interpreted and removed by a code cleaner pass. Before this fix, every `use` statement appeared to return false.

Now, when codecleaned input is empty, add an implicit 'return null' to the end. A null return value is consistent with, for example, the return value when declaring a function or class.

Fixes #259