Make reflecting commands superglobal-aware.
This only actually does something useful in `dump`. The rest of the reflecting commands require a variable target to be an object, which obviously none of the superglobals are.

This change makes `dump` actually do what you'd expect when you say `dump $GLOBALS` (or any other superglobals). Yay!

For everything but `dump`, this just makes it so the error message is sane, in case you do something like:

    >>> $_SERVER
    => (lots of output)
    >>> ls $_SERVER 
    InvalidArgumentException with message 'Unknown variable: $_SERVER'

Yeah, thanks PsySH. We both know you know about it. You just showed it to me.

After this change, the above will spit out:

    Unable to inspect a non-object

Which is a much more sane error message.