Don't show `=> null` for code with no return value
Prior to this change, all code input would print either a return value or an exception. This isn't technically correct, as non-expression statements (like `use` or `namespace` or a class definition) in PHP have no value.

But because all code input makes a round trip through `eval()`, and because in PHP `eval()` returns `null` if it has no explicit return value, everything in PsySH looked like it was returning `null`.

But!

This changes all that. We now implicitly return a NoReturnValue instance from all code input which doesn't have a return value, the same way we implicitly change the last expression in code input into a return statement.

One little caveat here is that if user code happens to return a NoReturnValue instance, all bets are off. That code will look like it has no return value. So don't do that.

In implementing this, we carve out two exceptions where we add implicit return values even though they're technically incorrect:

* when code input ends in an if/elseif/else statement
* when code input ends in a switch statement (to all non-passthrough cases)