Fix reference assignment regression.
In the refactor for v0.9.x we extracted the guts of the execution loop's closure, and reused the whole thing for arbitrary execution as well. But, in the process, we broke references. Since each iteration of the REPL happens in a fresh closure context, there was no way to maintain references across lines.

This change goes back to the v0.8.x style loop execution closure, where the iteration happens inside a single closure context. This way all reference assignments persist between lines of REPL input.

Unfortunately there are about 20 lines of duplicated code between the two types of execution context, but I don't see a way to avoid that while still keeping all the things that need to happen inside the closures where they belong.

Fixes #498