If there's an error initializing config... don't.
We're going to exit immediately anyway, so there's no reason to go through with creating a Shell instance. Make the "print the version" thing static, and bail without creating a Configuration or Shell instance if there was a problem.

This way if someone has an error in their config file it won't blow up multiple times, confusing everything in the process.

See #608