Fix requiring local files when running from a pre-built Phar.
PHP intercepts filesystem functions when called from inside a Phar, and resolves them to files with those names in the Phar, when they exist. This is normally what you want. That way when an Phar file tries to `require 'vendor/autoload.php'` it gets _the Phar's_ autoloader, not yours.

Unfortunately, it also seems to cache the result. Or maybe it only affects us because technically the code we're running in PsySH is being `eval`'d _inside_ the Phar? Not quite sure. Regardless, depending on where `require` is first called, it'll resolve to either the Phar's version or the filesystem version, and future calls will use that version even if they weren't called from inside the Phar.

To make it even more fun, even if you explicitly check what it'll be resolved to with `stream_resolve_include_path`, PHP omits the bit about it actually resolving to the `phar://` path when you try to run it :grimacing:

This is a hacky fix: implicitly convert `require 'vendor/autoload.php'` to `require './vendor/autoload.php'` for any filename that collides with something in the Phar. Filesystem calls to relative (or absolute) paths don't get hijacked by Phar.

Fixes #640