Add self-update option to replace the current version with the latest available version (#745)
This adds a --self-update option that will download the latest version from github and replace the currently installed version.

The default downloader is a very simple curl implementation, with a fallback to file_get/put_contents if curl is not installed, and will throw an exception if allow_url_fopen is not true. The download process will save the downloaded file to the system temp directory and extract the archive contents to the location of the current installed version.

The self-update option will keep the temporary backup copy of the current version in the system temp directory after updating.