Support Throwables directly in PsySH.
For historic reasons, we've been treating Errors differently than Exceptions, but we don't really need to? Since it's a backwards compatible change, just ... start supporting Throwables everywhere we previously supported Exceptions.

- Stop wrapping Error and TypeError in our own Exceptions.
- Deprecate `ErrorException::fromError` and `TypeErrorException::fromTypeError`.
- Support Throwable as $previous for all Exception classes.


Additionally, store the last Error in `$_e` / `lastException`. Technically we were already doing the equivalent of this, but it was the last Error wrapped in an ErrorException instance, which just made things more confusing. Now it stores Errors directly.

Fixes #750